refactor: move flowTransitionValidation into flows service
